Refactored Event data classes to be more readable
