+++ app/src/main/java/org/hackillinois/android/database/Converters.kt
@@ -3,7 +3,7 @@ package org.hackillinois.android.database
 import androidx.room.TypeConverter
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
-import org.hackillinois.android.database.entity.EventLocation
+import org.hackillinois.android.database.helpers.EventLocation
 
 class Converters {
     @TypeConverter
