+++ app/src/main/java/org/hackillinois/android/database/helpers/EventLocation.kt
@@ -1,0 +1,8 @@
+package org.hackillinois.android.database.helpers
+
+data class EventLocation(
+    val description: String,
+    val latitude: Double,
+    val longitude: Double,
+    val tags: List<String> = listOf(),
+)
