+++ app/src/main/java/org/hackillinois/android/repository/EventRepository.kt
@@ -6,6 +6,15 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import org.hackillinois.android.App
 import org.hackillinois.android.database.entity.*
+import org.hackillinois.android.model.event.AttendeeCheckInResponse
+import org.hackillinois.android.model.event.AttendeeData
+import org.hackillinois.android.model.event.EventCode
+import org.hackillinois.android.model.event.MeetingCheckInResponse
+import org.hackillinois.android.model.event.MeetingEventId
+import org.hackillinois.android.model.event.RSVPData
+import org.hackillinois.android.model.event.RegistrationData
+import org.hackillinois.android.model.event.StaffCheckInResponse
+import org.hackillinois.android.model.event.UserEventPair
 import org.json.JSONObject
 import retrofit2.HttpException
 
@@ -93,7 +102,9 @@ class EventRepository {
 
         suspend fun checkInAttendee(userToken: String, eventId: String): StaffCheckInResponse {
             val userTokenEventIdPair = UserEventPair(userToken, eventId)
-            var apiResponse = StaffCheckInResponse(-1, -1, "", RSVPData("", false, RegistrationData(AttendeeData(listOf()))))
+            var apiResponse = StaffCheckInResponse(-1, -1, "", RSVPData("", false, RegistrationData(
+                AttendeeData(listOf())
+            )))
 
             withContext(Dispatchers.IO) {
                 try {
